client to proxy conn build

+++ src/mysql/constants.rs
@@ -13,8 +13,8 @@ pub static DEFAULT_MAX_ALLOWED_PACKET: usize = 4 * 1024 * 1024;
 pub static SERVER_VERSION : &'static str = "1.0.0-sparrow";
 pub static MIN_PROTOCOL_VERSION : u8 = 10;
 
-pub static UTF8_GENERAL_CI: u16 = 33;
-pub static UTF8MB4_GENERAL_CI: u16 = 45;
+pub static UTF8_GENERAL_CI: u8 = 33;
+pub static UTF8MB4_GENERAL_CI: u8 = 45;
 
 bitflags! {
     /// MySql server status flags
